+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog
 
+## [1.1.0] - 04/04/2024
+
+- Arbre de fidelité : ajout d'un bouton pour faire un affichage compact des parts
+- corrections mineurs
+
 ## [1.0.0] - 01/04/2024
 
 - Version initial
 - Ajout d'un menu 'A propos'
 - Ajout d'un menu 'Historique'
 - Arbre de fidelité : ajout d'un bouton pour masquer de part à 100%
-- Arbre de fidelité : entête plus compacte
+- Arbre de fidelité : entête plus compact